# src/config/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Matching Back API
  version: 1.0.0
servers:
  - url: http://218.232.137.30:20080/api
  - url: http://localhost:8080/api
...
paths:
...
  /status:
    get:
      tags:
        - status
      summary: 현황 조회 (나의 현황, 내가 생성한 프로젝트 현황)
      description: 현황 조회 (나의 현황, 내가 생성한 프로젝트 현황)
      responses:
        '200':
          $ref: '#/components/responses/status200'
      security:
        - JWT: [ ]
  /status/user:
    get:
      tags:
        - status
      summary: 현황 조회 (나의 현황)
      description: 현황 조회 (나의 현황)
      responses:
        '200':
          $ref: '#/components/responses/statusUser200'
      security:
        - JWT: [ ]
  /status/myProject:
    get:
      tags:
        - status
      summary: 현황 조회 (내가 생성한 프로젝트 현황)
      description: 현황 조회 (내가 생성한 프로젝트 현황)
      responses:
        '200':
          $ref: '#/components/responses/statusProject200'
      security:
        - JWT: [ ]
  /status/user/{pjtSn}:
    get:
      tags:
        - status
      summary: 요청받은 프로젝트 디테일 정보
      description: 요청받은 프로젝트 디테일 정보
      parameters:
        - in: path
          name: pjtSn
          required: true
          description: 프로젝트 일련번호
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/myProject200'
        '400':
          description: 잘못된 요청
          content:
            text/plain:
              schema:
                type: string
                example: '요청된 프로젝트 조회 실패 - 해당 프로젝트가 없거나 권한이 없습니다: {에러메시지}'
      security:
        - JWT: [ ]
  /status/myProject/{pjtSn}/{userSn}:
    get:
      tags:
        - status
      summary: 프로젝트에 참여 요청중인 개발자 정보
      description: 프로젝트에 요청중인 개발자 정보
      parameters:
        - in: path
          name: pjtSn
          required: true
          description: 프로젝트 일련번호
          schema:
            type: integer
        - in: path
          name: userSn
          required: true
          description: 회원 일련번호
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/pfPfol200'
        '400':
          description: 잘못된 요청
          content:
            text/plain:
              schema:
                type: string
                example: '프로필 및 포트폴리오 조회 실패: {에러메시지}'
      security:
        - JWT: [ ]
  /status/{pjtSn}/req:
    post:
      tags:
        - status
      summary: 개발자에게 참여 요청 전송
      description: 개발자에게 참여 요청 전송
      parameters:
        - in: path
          name: pjtSn
          required: true
          description: 프로젝트 일련번호
          schema:
            type: integer
      requestBody:
        description: 회원 참여 요청 정보
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reqUser'
      responses:
        '200':
          description: 프로젝트 참여 요청 성공
          content:
            text:
              example: 프로젝트 참여 요청 성공
        '400':
          description: 잘못된 요청
          content:
            text/plain:
              schema:
                type: string
                example: '프로젝트 참여 요청 전송 중 에러 발생: {에러메세지}'
      security:
        - JWT: [ ]
  /status/{pjtSn}/req/{reqSn}:
    put:
      tags:
        - status
      summary: 개발자에게 참여 요청 상태 수정
      description: 개발자에게 참여 요청 수정
      parameters:
        - in: path
          name: pjtSn
          required: true
          description: 프로젝트 일련번호
          schema:
            type: integer
        - in: path
          name: reqSn
          required: true
          description: 요청 일련번호
          schema:
            type: integer
      requestBody:
        description: 회원 참여 요청 정보
        content:
          application/json:
            schema:
              type: object
              properties:
                reqStts:
                  type: string
                  description: 수정될 요청 상태
                  example: 'FINAL_REQ'
      responses:
        '201':
          description: 프로젝트 참여 상태 수정 성공
          content:
            text:
              example: 프로젝트 참여 상태 수정 성공
        '400':
          description: 잘못된 요청
          content:
            text/plain:
              schema:
                type: string
                example: '프로젝트 참여 요청 수정 중 에러 발생: {에러메세지}'
      security:
        - JWT: [ ]
...
components:
  schemas:
...
    reqUser:
      type: object
      properties:
        userSn:
          type: integer
          description: 회원 일련번호
          example: 1
        pjtRoleSn:
          type: integer
          description: 프로젝트 파트 일련번호
          example: 1
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    pfPfol200:
      description: 내 프트폴리오 조회 성공
      content:
        application/json:
          schema:
            type: object
            properties:
              PF_SN:
                type: integer
                description: 프로필 PK
                example: 1
              USER_SN:
                type: integer
                description: 회원 PK
                example: 1
              USER_NM:
                type: string
                description: 회원명
                example: 홍길동
              USER_IMG:
                type: string
                description: 회원 이미지
                example: www.www.com
              PF_INTRO:
                type: string
                description: 프로필 소개
                example: "OO의 프로필입니다."
              CAREER_NM:
                type: string
                description: 회사명
                example: 회사명
              ENTERING_DT:
                type: string
                description: 입사일
                example: 2023-01-01
              QUIT_DT:
                type: string
                description: 퇴사일
                example: 2023-12-31
              ST_NM:
                type: string
                description: 보유 스택
                example: REACT
              ST_LEVEL:
                type: string
                description: 보유 스택 수준
                example: HIGH
              INTEREST_NM:
                type: string
                description: 관심분야
                example: WEB
              URL_ADDR:
                type: string
                description: URL 주소
                example: http://github.com/user
              URL_INTRO:
                type: string
                description: URL 설명
                example: 깃허브
              PFOL_SN:
                type: integer
                description: 포트폴리오 Pk
                example: 1
              PFOL_NM:
                type: string
                description: 프로젝트명
                example: project1
              START_DT:
                type: string
                description: 프로젝트 시작일
                example: 2024-01-01
              END_DT:
                type: string
                description: 프로젝트 종료일
                example: 2024-05-01
              PERIOD:
                type: integer
                description: 프로젝트 진행 기간
                example: 121
              IMG:
                type: string
                description: 프로젝트 메인 이미지
                example: http://000.000.000/image1.jps
            example:
              profile:
                - PF_SN: 1
                  USER_SN: 1
                  USER_NM: 홍길동
                  USER_IMG: http://218.232.137.30:20090/matching/profile/img.jpg
                  career:
                    - CAREER_NM: 회사1
                      ENTERING_DT: 2023-01-01
                      QUIT_DT: 2023-10-31
                    - CAREER_NM: 회사2
                      ENTERING_DT: 2024-01-01
                      QUIT_DT: NULL
                  stack:
                    - ST_NM: JAVA
                      ST_LEVEL: HIGH
                    - ST_NM: REACT
                      ST_LEVEL: LOW
                  interest:
                    - INTEREST_NM: WEB
                    - INTEREST_NM: JAVA
                  url:
                    - URL_ADDR: www.www.com
                      URL_INTRO: 깃허브
              portfolioInfo:
                - PFOL_SN: 1
                  PFOL_NM: 프로젝트1
                  START_DT: 2023-01-01
                  END_DT: 2023-05-01
                  IMG: http://218.232.137.30:20090/matching/portfolio/portfolio.png
                  PERIOD: 121
                - PFOL_SN: 2
                  PFOL_NM: 프로젝트2
                  START_DT: 2024-07-01
                  END_DT: 2023-07-31
                  PERIOD: 31
                  stack:
                    - ST_NM: JAVA
                    - ST_NM: REACT
                  IMG: http://218.232.137.30:20090/matching/portfolio/portfolio.png
...
    myProject200:
      description: 내 프로젝트 조회 성공
      content:
        application/json:
          schema:
            type: object
            properties:
              pjtSn:
                type: integer
                description: 프로젝트 일련번호
                example: 1
              pjtNm:
                type: string
                description: 프로젝트명
                example: '프로젝트 관리 플랫폼'
              createdUserSn:
                description: 팀리더 회원 일련번호
                example: 3
              teamLeader:
                description: 팀리더
                example: "김동당"
              pjtImg:
                type: string
                description: 프로젝트 이미지 url
                example: 'https://example.com/project_management_platform.png'
              pjtStts:
                type: string
                description: 프로젝트 진행상태
                example: 모집중
              startDt:
                type: string
                description: 프로젝트 시작일
                example: 2024-07-20
              endDt:
                type: string
                description: 프로젝트 종료일
                example: 2024-09-20
              period:
                type: integer
                description: 프로젝트 기간
                example: 3
              durationUnit:
                type: string
                description: 프로젝트 기간(일/월)
                example: MONTH
              pjtIntro:
                type: string
                description: 프로젝트 소개
                example: '프로젝트 소개...'
              pjtDetail:
                type: string
                description: 프로젝트 상세 설명
                example: '프로젝트 상세 설명...'
              stack:
                type: string
                description: 사용 기술 스택
                example: 'Java, Node.Js'
              TO:
                type: integer
                description: 모집 인원
                example: 10
              PO:
                type: integer
                description: 현재 인원
                example: 3
              role:
                type: array
                description: 프로젝트 역할 및 인원
                items:
                  type: object
                  properties:
                    pjtRoleSn:
                      type: integer
                      description: 프로젝트 역할파트 일련번호
                      example: 2
                    part:
                      type: string
                      description: 역할 파트
                      example: '프런트엔드'
                    totalCnt:
                      type: integer
                      description: 총 인원 수
                      example: 5
                    cnt:
                      type: integer
                      description: 현재 참여 인원 수
                      example: 1
                    mem:
                      type: string
                      description: 참여 멤버
                      example: '김기, 박기'
              wanted:
                type: array
                description: 요구 경험
                items:
                  type: string
                  example: '앱, 웹'
          example:
            pjtSn: 1
            pjtNm: '프로젝트 관리 플랫폼'
            createdUserSn: 10
            teamLeader: "이세진"
            pjtImg: 'https://example.com/project_management_platform.png'
            pjtStts: '모집중'
            startDt: 2024-07-20
            endDt: null
            period: 3
            durationUnit: MONTH
            pjtIntro: '사용자들이 프로젝트를 효과적으로 관리할 수 있는 웹 플랫폼을 개발하는 프로젝트입니다.'
            pjtDetail: '이 프로젝트는 사용자들이 프로젝트를 효과적으로 관리할 수 있는 웹 플랫폼을 개발하는 것입니다. 프로젝트 생성 및 편집 기능, 일정 관리, 업무 할당, 팀원 협업 기능 등이 필요합니다.'
            stack: 'JavaScript,NODE.JS,REACT'
            TO: 10
            PO: 3
            role:
              - pjtRoleSn: 3
                part: '프런트엔드'
                totalCnt: 5
                cnt: 4
                mem: "홍길동"
              - pjtRoleSn: 4
                part: '백엔드'
                totalCnt: 4
                cnt: 0
                mem: "김신사, 박준"
            wanted: '웹 앱 개발 경험자, 프로젝트 관리 앱 경험자, 사용자 친화적인 UI/UX 디자인 능력'
...
    status200:
      description: 현황 조회 성공
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              properties:
                myReqList:
                  type: array
                  items:
                    type: object
                    properties:
                      REQ_SN:
                        type: integer
                        description: 요청 PK 값
                      PJT_SN:
                        type: integer
                        description: 프로젝트 PK 값
                      PJT_IMG:
                        type: string
                        description: 프로젝트 이미지
                      PJT_NM:
                        type: string
                        description: 프로젝트 이름
                      TOTAL_CNT:
                        type: integer
                        description: 총 모집 인원
                      PART:
                        type: string
                        description: 요청된 모집 분야
                      REQ_STTS:
                        type: string
                        description: 요청 상태코드
                      REQ_STTS_VAL:
                        type: string
                        description: 요청 상태코드 값
                projectReqList:
                  type: array
                  items:
                    type: object
                    properties:
                      PJT_SN:
                        type: integer
                        description: 프로젝트 PK 값
                      PJT_NM:
                        type: string
                        description: 프로젝트 이름
                      REQ_LIST:
                        type: array
                        items:
                          type: object
                          properties:
                            REQ_SN:
                              type: integer
                              description: 요청 PK 값
                            PF_SN:
                              type: integer
                              description: 프로필 PK 값
                            USER_SN:
                              type: integer
                              description: 요청 대상 회원 일련번호
                            USER_IMG:
                              type: string
                              description: 요청 대상 프로필 이미지
                            USER_NM:
                              type: string
                              description: 요청 대상 이름
                            PF_INTRO:
                              type: string
                              description: 요청 대상 프로필 설명
                            PART:
                              type: string
                              description: 요청된 모집 분야
                            REQ_STTS:
                              type: string
                              description: 요청 상태코드
                            REQ_STTS_VAL:
                              type: string
                              description: 요청 상태코드 값
          example:
            - myReqList:
              - REQ_SN: 1
                PJT_SN: 1
                PJT_IMG: http://218.232.137.30:20090/matching/project/img.png
                PJT_NM: project1
                TOTAL_CNT: 11
                PART: 프론트엔드
                REQ_STTS: REQ
                REQ_STTS_VAL: 요청
              - REQ_SN: 2
                PJT_SN: 2
                PJT_IMG: http://218.232.137.30:20090/matching/project/img.png
                PJT_NM: project2
                TOTAL_CNT: 7
                PART: 기획
                REQ_STTS: AGREE
                REQ_STTS_VAL: 수락
            - projectReqList:
                - PJT_SN: 3
                  PJT_NM: project3
                  REQ_LIST:
                    - REQ_SN: 3
                      PF_SN: 2
                      USER_SN: 1
                      USER_IMG: http://218.232.137.30:20090/matching/profile/img.png
                      USER_NM: 홍길동
                      PF_INTRO: 프로필 설명 1
                      PART: 백엔드
                      REQ_STTS: REJECTION
                      REQ_STTS_VAL: 거절
                    - REQ_SN: 4
                      PF_SN: 3
                      USER_SN: 2
                      USER_IMG: http://218.232.137.30:20090/matching/profile/img.png
                      USER_NM: 김철수
                      PF_INTRO: 프로필 설명 2
                      PART: 디자인
                      REQ_STTS: REQ
                      REQ_STTS_VAL: 요청
                - PJT_SN: 4
                  PJT_NM: project4
                  REQ_LIST:
                    - REQ_SN: 5
                      PF_SN: 5
                      USER_SN: 5
                      USER_IMG: http://218.232.137.30:20090/matching/profile/img.png
                      USER_NM: 김영희
                      PF_INTRO: 프로필 설명 3
                      PART: 프론트앤드
                      REQ_STTS: REQ
                    - REQ_SN: 6
                      PF_SN: 6
                      USER_SN: 6
                      USER_IMG: http://218.232.137.30:20090/matching/profile/img.png
                      USER_NM: 아무개
                      PF_INTRO: 프로필 설명 4
                      PART: 디자인
                      REQ_STTS: AGREE
                      REQ_STTS_VAL: 동의
    statusUser200:
      description: 개발자 기준 현황 조회 성공
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              properties:
                reqSn:
                  type: integer
                  description: 요청 PK 값
                pjtSn:
                  type: integer
                  description: 프로젝트 PK 값
                pjtImg:
                  type: string
                  description: 프로젝트 이미지
                pjtNm:
                  type: string
                  description: 프로젝트 이름
                TO:
                  type: integer
                  description: 총 모집 인원
                part:
                  type: string
                  description: 요청된 모집 분야
                reqSttsCd:
                  type: string
                  description: 요청 상태코드
                reqStts:
                  type: string
                  description: 요청 상태코드 값
          example:
            - reqSn: 1
              pjtSn: 1
              pjtImg: http://218.232.137.30:20090/matching/project/img.png
              pjtNm: project1
              TO: 11
              part: 프론트엔드
              reqSttsCd: REQ
              reqStts: 요청
            - reqSn: 2
              pjtSn: 2
              pjtImg: http://218.232.137.30:20090/matching/project/img.png
              pjtNm: project2
              TO: 6
              part: 기획
              reqSttsCd: AGREE
              reqStts: 1차 수락
    statusProject200:
      description: 프로젝트 기준 현황 조회 성공
      content:
        application/json:
          schema:
            type: array
            items:
              type: object
              properties:
                pjtSn:
                  type: integer
                  description: 프로젝트 PK 값
                pjtNm:
                  type: string
                  description: 프로젝트 이름
                reqList:
                  type: array
                  items:
                    type: object
                    properties:
                      reqSn:
                        type: integer
                        description: 요청 PK 값
                      userSn:
                        type: integer
                        description: 요청 대상 회원 일련번호
                      userNm:
                        type: string
                        description: 요청 대상 이름
                      userImg:
                        type: string
                        description: 요청 대상 프로필 이미지
                      part:
                        type: string
                        description: 요청된 모집 분야
                      reqStts:
                        type: string
                        description: 요청 상태코드 값
                      reqSttsCd:
                        type: string
                        description: 요청 상태코드
          example:
            - pjtSn: 3
              pjtNm: project3
              rqList:
              - reqSn: 3
                userSn: 1
                userNm: 홍길동
                userImg: http://218.232.137.30:20090/matching/profile/img.png
                part: 백엔드
                reqStts: 거절
                rqSttsCd: REJECTIOM
              - reqSn: 4
                userSn: 2
                userNm: 김철수
                userImg: http://218.232.137.30:20090/matching/profile/img.png
                part: 디자인
                reqStts: 요청
                rqSttsCd: REQ
            - pjtSn: 4
              pjtNm: project4
              reqList:
              - reqSn: 5
                userSn: 5
                userNm: 김영희
                userImg: http://218.232.137.30:20090/matching/profile/img.png
                part: 프론트앤드
                reqStts: REQ
                reqSttsCd: REQ
              - reqSn: 6
                userSn: 6
                userNm: 아무개
                userImg: http://218.232.137.30:20090/matching/profile/img.png
                part: 디자인
                reqStts: 1차 수락
                reqSttsCd: AGREE
...